Avoid conversion truncation/zeroextension on AMD64 by keeping bitcoiunt as uint.
enchmarkEncodeDigitsConstant1e4-4     156.76       158.75       1.01x
BenchmarkEncodeDigitsConstant1e5-4     167.12       168.81       1.01x
BenchmarkEncodeDigitsConstant1e6-4     165.45       165.59       1.00x

+++ flate/huffman_code.go
@@ -64,8 +64,8 @@ func (h hcode) code() (code uint16) {
 	return uint16(h)
 }
 
-func (h hcode) bits() (bits uint8) {
-	return uint8(h >> 16)
+func (h hcode) bits() (bits uint) {
+	return uint(h >> 16)
 }
 
 func maxNode() literalNode { return literalNode{math.MaxUint16, math.MaxInt32} }
